+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Latest
 
+- [Improvement] Add configurable SMTP authentication
 - [Improvement] Harmonize mysql username/password configuration parameters
 - [Feature] Configurable and pluggable data storage backends (#114)