+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Latest
 
+- [Bugfix] Fix `images pull all` command which failed on "all" image
 - [Improvement] Add configurable mongodb, SMTP and rabbitmq authentication
 - [Improvement] Harmonize mysql username/password configuration parameters
 - [Feature] Configurable and pluggable data storage backends (#114)