+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Latest
 
+- [Feature] Multiple platforms on a single server \o/
 - [Feature] Easily configure web proxy on the host
 - [Bugfix] Fix `images pull all` command which failed on "all" image
 - [Improvement] Add configurable mongodb, SMTP and rabbitmq authentication